+++ example/lib/main.dart
@@ -10,6 +10,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Side Sheet',
+      theme: ThemeData(primarySwatch: Colors.blueGrey),
       home: MyHomePage(title: 'Side Sheet Example'),
       debugShowCheckedModeBanner: false,
     );
@@ -25,6 +26,8 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  String text = 'no data';
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -44,7 +47,6 @@ class _MyHomePageState extends State<MyHomePage> {
                           context: context
                       ),
                   child: Text('OPEN RIGHT SIDE SHEET')),
-
               SizedBox(height: 10),
 
               ElevatedButton(
@@ -54,10 +56,55 @@ class _MyHomePageState extends State<MyHomePage> {
                           context: context
                       ),
                   child: Text('OPEN LEFT SIDE SHEET')),
+              SizedBox(height: 20),
+
+              ElevatedButton(
+                  onPressed: () async {
+                    final data = await SideSheet.left(
+                        body: bodyWithReturnArgs(context),
+                        context: context);
+
+                    setState(() {
+                      text = data;
+                    });
+                  },
+                  child: Text('OPEN LEFT SIDE SHEET WITH RETURN DATA')),
+              SizedBox(height: 10),
+
+              ElevatedButton(
+                  onPressed: () async {
+                    final data = await SideSheet.right(
+                        body: IconButton(
+                            icon: Icon(Icons.close),
+                            onPressed: () => Navigator.pop(context, 'Data oitrhatpoihotha')),
+                        context: context);
+
+                    setState(() {
+                      text = data;
+                    });
+                  },
+                  child: Text('OPEN RIGHT SIDE SHEET WITH RETURN DATA')),
+              SizedBox(height: 10),
+
+              Text('Arguments: $text')
             ],
           ),
         ),
       ),
     );
   }
+}
+
+Widget bodyWithReturnArgs(context) {
+  return Container(
+    alignment: Alignment.topCenter,
+    child: Column(
+      children: [
+        IconButton(
+            icon: Icon(Icons.close),
+            onPressed: () => Navigator.pop(context, 'Data Returns')),
+        Text('Body')
+      ],
+    ),
+  );
 }