+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+* Data returns when close sheet.
+* Accessibility labels
+* Choose dismissible sheet when click out of area
+
 ## 0.0.3
 
 * Change class SideSheet.